Fix a warning in a printf statement.

+++ src/resmom/mom_job_func.c
@@ -464,7 +464,7 @@ int conn_qsub(
     snprintf(EMsg, 1024, "Failed to connect to %s at address %s:%d",
              hostname,
              remote_addr,
-             port);
+             (int)port);
 
     return(-1);
     }
